:recycle: split files
