+++ main.go
@@ -10,65 +10,24 @@ void greeter_signal_connect(LightDMGreeter* greeter);
 import "C"
 
 import (
-	"encoding/json"
 	"fmt"
 	"log"
-	"math/rand"
-	"os"
-	"time"
 	"unsafe"
-
-	"github.com/gotk3/gotk3/gdk"
-	"github.com/gotk3/gotk3/gtk"
 )
 
-// CONSTANTS
+/*************/
+/* CONSTANTS */
+/*************/
+
 const BASE_PATH = "/etc/lightdm/lightdm-micro-greeter/"
 const CONFIG_FILE = BASE_PATH + "config.json"
 
-// GLOBAL VARS
-// TODO this is ulgy, can I avoid them ?
+/***************/
+/* GLOBAL VARS */
+/***************/
+
+// flag, nil in multi user mode
 var c_username *C.char = nil
-var entry *gtk.Entry
-var label *gtk.Label
-
-/**********************/
-/* JSON Configuration */
-/**********************/
-type Configuration struct {
-	Username  string
-	Wallpaper string
-	Entry     struct {
-		WidthChars     int
-		Margin         int
-		XLocationRatio float32
-		YLocationRatio float32
-	}
-}
-
-func loadConfig() (config Configuration, err error) {
-	config.Username = ""
-	config.Wallpaper = ""
-	config.Entry.WidthChars = 10
-	config.Entry.Margin = 10
-	config.Entry.XLocationRatio = 0.5
-	config.Entry.YLocationRatio = 0.5
-
-	file, err := os.Open(CONFIG_FILE)
-	if err != nil {
-		log.Print("[load_config] error opening " + CONFIG_FILE)
-		return
-	}
-	defer file.Close()
-
-	decoder := json.NewDecoder(file)
-	err = decoder.Decode(&config)
-	if err != nil {
-		log.Print("[load_config] " + CONFIG_FILE + "is not a valid JSON")
-	}
-
-	return
-}
 
 /*****************************/
 /* Light DM Server Callbacks */
@@ -103,9 +62,9 @@ func show_prompt_cb(greeter *C.LightDMGreeter, text *C.char, prompt_type C.Light
 	}
 }
 
-/*************/
-/* Utilities */
-/*************/
+/************************/
+/* GTK Callback Factory */
+/************************/
 
 func createEntryCallback(greeter *C.LightDMGreeter) func() {
 	return func() {
@@ -116,7 +75,8 @@ func createEntryCallback(greeter *C.LightDMGreeter) func() {
 		defer C.free(unsafe.Pointer(c_input))
 
 		if C.lightdm_greeter_get_is_authenticated(greeter) != 0 {
-			// start_session ?
+			// session starting
+			label.SetText("session starts...")
 			log.Print("[entry_callback] authentication ok")
 		} else if C.lightdm_greeter_get_in_authentication(greeter) != 0 {
 			// give pwd
@@ -131,32 +91,12 @@ func createEntryCallback(greeter *C.LightDMGreeter) func() {
 	}
 }
 
-func loadWallpaper(fpath string, width, height int) (bg *gtk.Image, err error) {
-	filestat, err := os.Stat(fpath)
-	if err != nil {
-		log.Print("[load_wallpaper] error opening " + fpath)
-		return
-	}
-	if filestat.IsDir() {
-		files, _ := os.ReadDir(fpath)
-		rand.Seed(time.Now().UnixNano())
-		fpath += files[rand.Intn(len(files))].Name()
-		log.Print("[load_wallpaper] randomly picking " + fpath)
-	}
-	pixbuf, err := gdk.PixbufNewFromFileAtSize(fpath, width, height)
-	if err != nil {
-		log.Print("[load_wallpaper] error loading " + fpath)
-		return
-	}
-	bg, err = gtk.ImageNewFromPixbuf(pixbuf)
-	return
-}
+/*************************/
+/* lightdm-micro-greeter */
+/*************************/
 
 func initGreeter() (greeter *C.LightDMGreeter, err error) {
 	greeter = C.lightdm_greeter_new()
-	// TODO fix invalid pointer at runtime
-	// defer C.free(unsafe.Pointer(greeter))
-	// should I really free this ? Does glib free it for me ? should I use g_free ?
 	if C.lightdm_greeter_connect_to_daemon_sync(greeter, nil) == 0 {
 		log.Print("[init_greeter] can't connect to LightDM deamon")
 		err = fmt.Errorf("can't connect to LightDM deamon")
@@ -167,70 +107,11 @@ func initGreeter() (greeter *C.LightDMGreeter, err error) {
 	return
 }
 
-func initWindow() (window *gtk.Window, width, height int, err error) {
-	window, err = gtk.WindowNew(gtk.WINDOW_TOPLEVEL)
-	if err != nil {
-		return
-	}
-	// seems useless, never called when session starts
-	window.Connect("destroy", func() {
-		log.Print("destroy signal called: quitting gtk")
-		gtk.MainQuit()
-	})
-
-	// get screen information to resize as full screen
-	// .Fullscreen() is not working here
-	display, err := window.GetDisplay()
-	if err != nil {
-		return
-	}
-	monitor, err := display.GetPrimaryMonitor()
-	if err != nil {
-		return
-	}
-	rect := monitor.GetGeometry()
-	width = rect.GetWidth()
-	height = rect.GetHeight()
-
-	window.Resize(width, height)
-
-	return
-}
-
-func initEntryBox(margin, widthChars int, callback func()) (box *gtk.Box, err error) {
-	box, err = gtk.BoxNew(gtk.ORIENTATION_VERTICAL, margin)
-	if err != nil {
-		return
-	}
-	box.SetHAlign(gtk.ALIGN_CENTER)
-	box.SetVAlign(gtk.ALIGN_CENTER)
-
-	label, err = gtk.LabelNew("username")
-	if err != nil {
-		return
-	}
-	label.SetHAlign(gtk.ALIGN_CENTER)
-	label.SetVAlign(gtk.ALIGN_CENTER)
-	box.Add(label)
-
-	entry, err = gtk.EntryNew()
-	if err != nil {
-		return
-	}
-	entry.SetHAlign(gtk.ALIGN_CENTER)
-	entry.SetVAlign(gtk.ALIGN_CENTER)
-	entry.SetWidthChars(widthChars)
-	entry.Connect("activate", callback)
-	box.Add(entry)
-
-	return
-}
-
 func main() {
 	log.Print("[main] start up")
 
 	// Reading configuration file
-	config, _ := loadConfig()
+	config, _ := loadConfig(CONFIG_FILE)
 
 	// Start greeter
 	greeter, err := initGreeter()
@@ -238,48 +119,7 @@ func main() {
 		log.Fatalf("[start_greeter] fatal error: %s", err)
 	}
 
-	// Init GUI
-	log.Print("[main] gtk init")
-	gtk.Init(nil)
-
-	// fullscreen window
-	win, width, height, err := initWindow()
-	if err != nil {
-		log.Fatalf("[init_window] fatal error: %s", err)
-	}
-
-	// simple fixed layout
-	layout, _ := gtk.FixedNew()
-	win.Add(layout)
-
-	// set background image, auto scaling while preserving aspect ratio
-	bg, err := loadWallpaper(BASE_PATH+config.Wallpaper, width, height)
-	if err != nil {
-		log.Print("[load_wallpaper] default white screen")
-	} else {
-		log.Print("[load_wallpaper] wallpaper loaded")
-		layout.Put(bg, 0, 0)
-	}
-
-	// init entry box
-	box, _ := initEntryBox(
-		config.Entry.Margin, config.Entry.WidthChars,
-		createEntryCallback(greeter))
-	/* TODO find a cleaner way to acheive this, might induce flickering
-	   for now, I have to put the box and render it before having access to its size */
-	// set box centered
-	layout.Add(box)
-	win.ShowAll()
-	// now that size is known, compute center
-	center_x := int(float32(width) * config.Entry.XLocationRatio)
-	center_y := int(float32(height) * config.Entry.YLocationRatio)
-	offset_x := box.GetAllocatedWidth() / 2
-	offset_y := box.GetAllocatedWidth() / 2
-	// center box
-	layout.Remove(box)
-	layout.Put(box, center_x-offset_x, center_y-offset_y)
-	// set cursor in entry
-	entry.GrabFocus()
+	initUI(config, createEntryCallback(greeter))
 
 	// Starts autologin if provided
 	if config.Username != "" {
@@ -288,6 +128,4 @@ func main() {
 		C.lightdm_greeter_authenticate(greeter, c_username, nil)
 	}
 
-	log.Print("[main] gtk start main loop")
-	gtk.Main()
 }
